fs_bsp_to_xi_mapping gets its own file.
