+++ src/libxively/sft/xi_sft_logic.c
@@ -15,6 +15,8 @@
 #include <xi_bsp_io_fs.h>
 #include <xi_bsp_fwu.h>
 
+#include <xi_fs_bsp_to_xi_mapping.h>
+
 xi_state_t xi_sft_make_context( xi_sft_context_t** context,
                                 const char** updateable_files,
                                 uint16_t updateable_files_count,
